// src/app/printers/interfaces/product.interface.ts
export interface Printer {
  id?: string;
  marca: string;
  modelo: string;
  descrip_corta: string;
  descripcion: string;
  alt_img?: string;
  tm_calentamiento: string;
  v_p_impresion: string;
  v_impresion_c: string;
  memoria?: string;
  disco?: string;
  dimensiones: string;
  peso: string;
}